+++ XFGoogleMapSample/iOS/AppDelegate.cs
@@ -4,6 +4,7 @@ using System.Linq;
 
 using Foundation;
 using UIKit;
+using Xamarin.Forms.GoogleMaps.iOS;
 
 namespace XFGoogleMapSample.iOS
 {
@@ -13,7 +14,14 @@ namespace XFGoogleMapSample.iOS
         public override bool FinishedLaunching(UIApplication app, NSDictionary options)
         {
             global::Xamarin.Forms.Forms.Init();
-            Xamarin.FormsGoogleMaps.Init(Variables.GOOGLE_MAPS_IOS_API_KEY); 
+
+            // Override default ImageFactory by your implementation. 
+            var platformConfig = new PlatformConfig
+            {
+                ImageFactory = new CachingImageFactory()
+            };
+
+            Xamarin.FormsGoogleMaps.Init(Variables.GOOGLE_MAPS_IOS_API_KEY, platformConfig); 
             LoadApplication(new App());
 
             return base.FinishedLaunching(app, options);