+++ Xamarin.Forms.GoogleMaps/Xamarin.Forms.GoogleMaps.Android/Extensions/ActivityExtensions.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using Android.App;
+using Android.Util;
+
+namespace Xamarin.Forms.GoogleMaps.Android.Extensions
+{
+    internal static class ActivityExtensions
+    {
+        public static float GetScaledDensity(this Activity self) 
+        {
+            var metrics = new DisplayMetrics();
+            self.WindowManager.DefaultDisplay.GetMetrics(metrics);
+            return metrics.ScaledDensity;
+        }
+    }
+}