+++ Xamarin.Forms.GoogleMaps/Xamarin.Forms.GoogleMaps.Android/MapRenderer.cs
@@ -7,8 +7,8 @@ using Android.OS;
 using Java.Lang;
 using Xamarin.Forms.Platform.Android;
 using Math = System.Math;
-using Android.Util;
 using Android.App;
+using Android.Content;
 using Android.Graphics;
 using Xamarin.Forms.GoogleMaps.Logics.Android;
 using Xamarin.Forms.GoogleMaps.Logics;
@@ -20,7 +20,7 @@ using Xamarin.Forms.GoogleMaps.Internals;
 
 namespace Xamarin.Forms.GoogleMaps.Android
 {
-    public class MapRenderer : ViewRenderer,
+    public class MapRenderer : ViewRenderer<Map, global::Android.Views.View>,
         GoogleMap.IOnMapClickListener,
         GoogleMap.IOnMapLongClickListener,
         GoogleMap.IOnMyLocationButtonClickListener
@@ -29,7 +29,7 @@ namespace Xamarin.Forms.GoogleMaps.Android
         readonly UiSettingsLogic _uiSettingsLogic = new UiSettingsLogic();
         readonly BaseLogic<GoogleMap>[] _logics;
 
-        public MapRenderer() : base()
+        public MapRenderer(Context context) : base(context)
         {
             _cameraLogic = new CameraLogic(UpdateVisibleRegion);
 
@@ -39,35 +39,35 @@ namespace Xamarin.Forms.GoogleMaps.Android
                 new PolylineLogic(),
                 new PolygonLogic(),
                 new CircleLogic(),
-                new PinLogic(OnMarkerCreating, OnMarkerCreated, OnMarkerDeleting, OnMarkerDeleted),
+                new PinLogic(context, Config.BitmapDescriptorFactory, 
+                    OnMarkerCreating, OnMarkerCreated, OnMarkerDeleting, OnMarkerDeleted),
                 new TileLayerLogic(),
-                new GroundOverlayLogic()
+                new GroundOverlayLogic(Config.BitmapDescriptorFactory)
             };
         }
-
-        public MapRenderer(IntPtr javaReference, global::Android.Runtime.JniHandleOwnership transfer) : this() { }
 
         static Bundle s_bundle;
         internal static Bundle Bundle { set { s_bundle = value; } }
 
+        internal static PlatformConfig Config { private get; set; }
+
+        // ReSharper disable once MemberCanBePrivate.Global
         protected GoogleMap NativeMap { get; private set; }
 
-        protected Map Map => (Map)Element;
+        // ReSharper disable once MemberCanBePrivate.Global
+        protected Map Map => Element;
 
-        private GoogleMap _oldNativeMap;
-        private Map _oldMap;
+        private bool _ready = false;
+        private bool _onLayout = false;
 
-        bool _ready = false;
-        bool _onLayout = false;
-
-        float _scaledDensity;
+        private float _scaledDensity;
 
         public override SizeRequest GetDesiredSize(int widthConstraint, int heightConstraint)
         {
             return new SizeRequest(new Size(Context.ToPixels(40), Context.ToPixels(40)));
         }
 
-        protected override async void OnElementChanged(ElementChangedEventArgs<View> e)
+        protected override async void OnElementChanged(ElementChangedEventArgs<Map> e)
         {
             base.OnElementChanged(e);
 
@@ -90,7 +90,31 @@ namespace Xamarin.Forms.GoogleMaps.Android
                 return;
             }
 
-            var oldMapView = (MapView)Control;
+            // Uninitialize old view
+            GoogleMap oldNativeMap = null;
+            Map oldMap = null;
+            if (e.OldElement != null)
+            {
+                try
+                {
+                    var oldNativeView = Control as MapView;
+                    // ReSharper disable once PossibleNullReferenceException
+                    oldNativeMap = await oldNativeView?.GetGoogleMapAsync();
+                    oldMap = e.OldElement;
+                    Uninitialize(oldNativeMap, oldMap);
+                    oldNativeView?.Dispose();
+                } 
+                catch (System.Exception ex) 
+                {
+                    var message = ex.Message;
+                    System.Diagnostics.Debug.WriteLine($"Uninitialize old view failed. - {message}");
+                }
+            }
+
+            if (e.NewElement == null)
+            {
+                return;
+            }
 
             var mapView = new MapView(Context);
             mapView.OnCreate(s_bundle);
@@ -100,39 +124,24 @@ namespace Xamarin.Forms.GoogleMaps.Android
             var activity = Context as Activity;
             if (activity != null)
             {
-                var metrics = new DisplayMetrics();
-                activity.WindowManager.DefaultDisplay.GetMetrics(metrics);
-                _scaledDensity = metrics.ScaledDensity;
+                _scaledDensity = activity.GetScaledDensity();
+                _cameraLogic.ScaledDensity = _scaledDensity;
                 foreach (var logic in _logics)
                 {
                     logic.ScaledDensity = _scaledDensity;
                 }
-
-                _cameraLogic.ScaledDensity = _scaledDensity;
             }
 
-            if (e.OldElement != null)
+            var newMap = e.NewElement;
+            NativeMap = await mapView.GetGoogleMapAsync();
+
+            foreach (var logic in _logics)
             {
-                var oldMapModel = (Map)e.OldElement;
-
-                oldMapModel.OnSnapshot -= OnSnapshot;
-                _cameraLogic.Unregister();
+                logic.Register(oldNativeMap, oldMap, NativeMap, newMap);
+                logic.ScaledDensity = _scaledDensity;
             }
 
-            if (oldMapView != null)
-            {
-                _oldNativeMap = await oldMapView.GetGoogleMapAsync();
-                _oldMap = (Map)e.OldElement;
-            }
-
-            NativeMap = await ((MapView)Control).GetGoogleMapAsync();
-
-            _cameraLogic.Register(Map, NativeMap);
-            Map.OnSnapshot += OnSnapshot;
-
-            _uiSettingsLogic.Register(Map, NativeMap);
-
-            OnMapReady(NativeMap);
+            OnMapReady(NativeMap, newMap);
         }
 
         private void OnSnapshot(TakeSnapshotMessage snapshotMessage)
@@ -146,15 +155,18 @@ namespace Xamarin.Forms.GoogleMaps.Android
             }));
         }
 
-        void OnMapReady(GoogleMap map)
+        protected virtual void OnMapReady(GoogleMap nativeMap, Map map)
         {
-            if (map != null)
+            if (nativeMap != null)
             {
-                _cameraLogic.Register(Map, NativeMap);
+                _cameraLogic.Register(map, nativeMap);
 
-                map.SetOnMapClickListener(this);
-                map.SetOnMapLongClickListener(this);
-                map.SetOnMyLocationButtonClickListener(this);
+                _uiSettingsLogic.Register(map, nativeMap);
+                Map.OnSnapshot += OnSnapshot;
+
+                nativeMap.SetOnMapClickListener(this);
+                nativeMap.SetOnMapLongClickListener(this);
+                nativeMap.SetOnMyLocationButtonClickListener(this);
 
                 UpdateIsShowingUser(_uiSettingsLogic.MyLocationButtonEnabled);
                 UpdateHasScrollEnabled(_uiSettingsLogic.ScrollGesturesEnabled);
@@ -168,11 +180,6 @@ namespace Xamarin.Forms.GoogleMaps.Android
 
                 SetMapType();
                 SetPadding();
-            }
-
-            foreach (var logic in _logics)
-            {
-                logic.Register(_oldNativeMap, _oldMap, NativeMap, Map);
             }
 
             _ready = true;
@@ -201,14 +208,12 @@ namespace Xamarin.Forms.GoogleMaps.Android
             else if (changed && NativeMap != null)
             {
                 UpdateVisibleRegion(NativeMap.CameraPosition.Target);
-                _cameraLogic.MoveCamera(Map.InitialCameraUpdate);
             }
         }
 
         void InitializeLogic()
         {
             _cameraLogic.MoveCamera(Map.InitialCameraUpdate);
-            //_cameraLogic.MoveToRegion(((Map)Element).LastMoveToRegion, false);
 
             foreach (var logic in _logics)
             {
@@ -291,8 +296,10 @@ namespace Xamarin.Forms.GoogleMaps.Android
 
         private void UpdateIsShowingUser(bool? initialMyLocationButtonEnabled = null)
         {
+#pragma warning disable 618
             NativeMap.MyLocationEnabled = Map.IsShowingUser;
             NativeMap.UiSettings.MyLocationButtonEnabled = initialMyLocationButtonEnabled ?? Map.IsShowingUser;
+#pragma warning restore 618
         }
 
         private void UpdateMyLocationEnabled()
@@ -302,20 +309,26 @@ namespace Xamarin.Forms.GoogleMaps.Android
 
         private void UpdateHasScrollEnabled(bool? initialScrollGesturesEnabled = null)
         {
+#pragma warning disable 618
             NativeMap.UiSettings.ScrollGesturesEnabled = initialScrollGesturesEnabled ?? Map.HasScrollEnabled;
+#pragma warning restore 618
         }
 
         private void UpdateHasZoomEnabled(
             bool? initialZoomControlsEnabled = null, 
             bool? initialZoomGesturesEnabled = null)
         {
+#pragma warning disable 618
             NativeMap.UiSettings.ZoomControlsEnabled = initialZoomControlsEnabled ?? Map.HasZoomEnabled;
             NativeMap.UiSettings.ZoomGesturesEnabled = initialZoomGesturesEnabled ?? Map.HasZoomEnabled;
+#pragma warning restore 618
         }
 
         private void UpdateHasRotationEnabled(bool? initialRotateGesturesEnabled = null)
         {
+#pragma warning disable 618
             NativeMap.UiSettings.RotateGesturesEnabled = initialRotateGesturesEnabled ?? Map.HasRotationEnabled;
+#pragma warning restore 618
         }
 
         private void UpdateIsTrafficEnabled()
@@ -401,7 +414,8 @@ namespace Xamarin.Forms.GoogleMaps.Android
             var lr = projection.FromScreenLocation(new global::Android.Graphics.Point(width, height));
             var dlat = Math.Max(Math.Abs(ul.Latitude - lr.Latitude), Math.Abs(ur.Latitude - ll.Latitude));
             var dlong = Math.Max(Math.Abs(ul.Longitude - lr.Longitude), Math.Abs(ur.Longitude - ll.Longitude));
-            ((Map)Element).VisibleRegion = new MapSpan(
+#pragma warning disable 618
+            Element.VisibleRegion = new MapSpan(
                     new Position(
                         pos.Latitude,
                         pos.Longitude
@@ -409,6 +423,8 @@ namespace Xamarin.Forms.GoogleMaps.Android
                 dlat,
                 dlong
             );
+#pragma warning restore 618
+            Element.Region = projection.VisibleRegion.ToRegion();
         }
 
         #region Overridable Members
@@ -455,30 +471,61 @@ namespace Xamarin.Forms.GoogleMaps.Android
 
         #endregion
 
+        private void Uninitialize(GoogleMap nativeMap, Map map) 
+        {
+            try
+            {
+                if (nativeMap == null)
+                {
+                    System.Diagnostics.Debug.WriteLine($"Uninitialize failed - {nameof(nativeMap)} is null");
+                    return;
+                }
+
+                if (map == null)
+                {
+                    System.Diagnostics.Debug.WriteLine($"Uninitialize failed - {nameof(map)} is null");
+                    return;
+                }
+
+                _uiSettingsLogic.Unregister();
+
+                map.OnSnapshot -= OnSnapshot;
+                _cameraLogic.Unregister();
+
+                foreach (var logic in _logics)
+                {
+                    logic.Unregister(nativeMap, map);
+                }
+
+                nativeMap.SetOnMapClickListener(null);
+                nativeMap.SetOnMapLongClickListener(null);
+                nativeMap.SetOnMyLocationButtonClickListener(null);
+
+                nativeMap.MyLocationEnabled = false;
+                nativeMap.Dispose();
+            }
+            catch (System.Exception ex)
+            {
+                var message = ex.Message;
+                System.Diagnostics.Debug.WriteLine($"Uninitialize failed. - {message}");
+            }
+        }
+
         bool _disposed;
         protected override void Dispose(bool disposing)
         {
             if (disposing && !_disposed)
             {
                 _disposed = true;
-
-                _uiSettingsLogic.Unregister();
-
-                Map.OnSnapshot -= OnSnapshot;
-                _cameraLogic.Unregister();
-
-                foreach (var logic in _logics)
-                    logic.Unregister(NativeMap, Map);
+                Uninitialize(NativeMap, Map);
 
                 if (NativeMap != null)
                 {
-                    NativeMap.SetOnMapClickListener(null);
-                    NativeMap.SetOnMapLongClickListener(null);
-                    NativeMap.SetOnMyLocationButtonClickListener(null);
-
-                    NativeMap.MyLocationEnabled = false;
                     NativeMap.Dispose();
+                    NativeMap = null;
                 }
+
+                (Control as MapView)?.OnDestroy();
             }
 
             base.Dispose(disposing);