+++ Xamarin.Forms.GoogleMaps/Xamarin.Forms.GoogleMaps.iOS/Factories/IImageFactory.cs
@@ -1,0 +1,9 @@
+﻿using UIKit;
+
+namespace Xamarin.Forms.GoogleMaps.iOS.Factories
+{
+    public interface IImageFactory
+    {
+        UIImage ToUIImage(BitmapDescriptor descriptor);
+    }
+}