+++ Xamarin.Forms.GoogleMaps/Xamarin.Forms.GoogleMaps.iOS/PlatformConfig.cs
@@ -1,0 +1,9 @@
+﻿using Xamarin.Forms.GoogleMaps.iOS.Factories;
+
+namespace Xamarin.Forms.GoogleMaps.iOS
+{
+    public sealed class PlatformConfig
+    {
+        public IImageFactory ImageFactory { get; set; }   
+    }
+}